combat fixes
