+++ tuxemon/core/states/combat/__init__.py
@@ -1,4 +1,5 @@
 from combat import CombatState
+from combat import WaitForInputState
 from combat_menus import MainCombatMenuState
 from combat_menus import TechniqueMenuState
 from combat_animations import CombatAnimations
