+++ tuxemon/core/states/combat/combat_menus.py
@@ -7,7 +7,7 @@ from functools import partial
 from core.components.menu import PopUpMenu
 from core.components.menu.interface import MenuItem
 from core.components.menu.menu import Menu
-from core.components.monster import Technique
+from core.components.technique import Technique
 
 # Create a logger for optional handling of debug messages.
 logger = logging.getLogger(__name__)
