--- tuxemon/resources/db/status/ailment.json
@@ -1,4 +1,0 @@
-{
-  "name": "ailment",
-  "icon": "gfx/ui/icons/party/party_icon02.png"
-}
