--- tuxemon/resources/db/status/fainted.json
@@ -1,4 +1,0 @@
-{
-  "name": "fainted",
-  "icon": "gfx/ui/icons/party/party_icon03.png"
-}
