--- tuxemon/resources/db/status/normal.json
@@ -1,4 +1,0 @@
-{
-  "name": "normal",
-  "icon": ""
-}
