--- tuxemon/resources/db/status/poisoned.json
@@ -1,4 +1,0 @@
-{
-  "name": "poisoned",
-  "icon": "gfx/ui/icons/status/poison-icon.png"
-}
