+++ tuxemon/resources/db/technique/poison_sting.json
@@ -1,14 +1,15 @@
 {
     "animation": "greendrip",
     "sfx": "bite1.ogg",
+    "icon": "",
     "category": "special",
     "effects": [
         "damage",
         "poison"
-    ], 
-    "id": 2, 
-    "name": "Poison Sting", 
-    "power": 40, 
+    ],
+    "id": 2,
+    "name": "Poison Sting",
+    "power": 40,
     "types": [
         "Poison"
     ]
