+++ tuxemon/resources/db/technique/pound.json
@@ -1,13 +1,14 @@
 {
     "animation": "pound",
     "sfx": "blaster1.ogg",
+    "icon": "",
     "category": "physical",
     "effects": [
         "damage"
-    ], 
-    "id": 1, 
-    "name": "Pound", 
-    "power": 50, 
+    ],
+    "id": 1,
+    "name": "Pound",
+    "power": 50,
     "types": [
         "Normal"
     ]
