+++ tuxemon/resources/db/technique/run.json
@@ -1,6 +1,7 @@
 {
     "animation": "",
     "sfx": "blaster1.ogg",
+    "icon": "",
     "category": "",
     "effects": [],
     "id": 6,
