+++ tuxemon/resources/db/technique/status_fainted.json
@@ -1,6 +1,7 @@
 {
   "animation": "",
   "sfx": "faint.ogg",
+  "icon": "",
   "category": "faint",
   "effects": [],
   "id": 7,
