+++ tuxemon/resources/db/technique/swap.json
@@ -1,6 +1,7 @@
 {
     "animation": "teeth",
     "sfx": "blaster1.ogg",
+    "icon": "",
     "category": "meta",
     "effects": ["swap"],
     "id": 8,
