+++ tuxemon/resources/db/technique/water_shot.json
@@ -1,13 +1,14 @@
 {
     "animation": "watershot",
     "sfx": "blaster1.ogg",
+    "icon": "",
     "category": "physical",
     "effects": [
         "damage"
-    ], 
-    "id": 3, 
-    "name": "Water Shot", 
-    "power": 50, 
+    ],
+    "id": 3,
+    "name": "Water Shot",
+    "power": 50,
     "types": [
         "Water"
     ]
